refactor: put route error logic into error boundary component
